Move filter cutoff buffer transform into a separate utility module.
